# Precache dictionary hashes