--- a/src/dictionary.c
+++ b/src/dictionary.c
@@ -54,7 +54,7 @@ void dict_free(dictionary* d) {
   // no cleanup currently needed
 }
 
-uint32_t dict_hash_scalar(dictionary* d, dictionary* x, R_len_t i) {
+uint32_t dict_hash_with(dictionary* d, dictionary* x, R_len_t i) {
   uint32_t hash = x->hash[i];
 
   // Quadratic probing: will try every slot if d->size is power of 2
@@ -85,6 +85,11 @@ uint32_t dict_hash_scalar(dictionary* d, dictionary* x, R_len_t i) {
   Rf_errorcall(R_NilValue, "Internal error: Dictionary is full!");
 }
 
+uint32_t dict_hash_scalar(dictionary* d, R_len_t i) {
+  return dict_hash_with(d, d, i);
+}
+
+
 void dict_put(dictionary* d, uint32_t hash, R_len_t i) {
   d->key[hash] = i;
   d->used++;
@@ -103,7 +108,7 @@ SEXP vctrs_unique_loc(SEXP x) {
 
   R_len_t n = vec_size(x);
   for (int i = 0; i < n; ++i) {
-    uint32_t hash = dict_hash_scalar(&d, &d, i);
+    uint32_t hash = dict_hash_scalar(&d, i);
 
     if (d.key[hash] == DICT_EMPTY) {
       dict_put(&d, hash, i);
@@ -125,7 +130,7 @@ SEXP vctrs_duplicated_any(SEXP x) {
   R_len_t n = vec_size(x);
 
   for (int i = 0; i < n; ++i) {
-    uint32_t hash = dict_hash_scalar(&d, &d, i);
+    uint32_t hash = dict_hash_scalar(&d, i);
 
     if (d.key[hash] == DICT_EMPTY) {
       dict_put(&d, hash, i);
@@ -145,7 +150,7 @@ SEXP vctrs_n_distinct(SEXP x) {
 
   R_len_t n = vec_size(x);
   for (int i = 0; i < n; ++i) {
-    uint32_t hash = dict_hash_scalar(&d, &d, i);
+    uint32_t hash = dict_hash_scalar(&d, i);
 
     if (d.key[hash] == DICT_EMPTY)
       dict_put(&d, hash, i);
@@ -164,7 +169,7 @@ SEXP vctrs_id(SEXP x) {
   int* p_out = INTEGER(out);
 
   for (int i = 0; i < n; ++i) {
-    uint32_t hash = dict_hash_scalar(&d, &d, i);
+    uint32_t hash = dict_hash_scalar(&d, i);
 
     if (d.key[hash] == DICT_EMPTY) {
       dict_put(&d, hash, i);
@@ -184,7 +189,7 @@ SEXP vctrs_match(SEXP needles, SEXP haystack) {
   // Load dictionary with haystack
   R_len_t n_haystack = vec_size(haystack);
   for (int i = 0; i < n_haystack; ++i) {
-    uint32_t hash = dict_hash_scalar(&d, &d, i);
+    uint32_t hash = dict_hash_scalar(&d, i);
 
     if (d.key[hash] == DICT_EMPTY) {
       dict_put(&d, hash, i);
@@ -200,7 +205,7 @@ SEXP vctrs_match(SEXP needles, SEXP haystack) {
   int* p_out = INTEGER(out);
 
   for (int i = 0; i < n_needle; ++i) {
-    uint32_t hash = dict_hash_scalar(&d, &d_needles, i);
+    uint32_t hash = dict_hash_with(&d, &d_needles, i);
     if (d.key[hash] == DICT_EMPTY) {
       p_out[i] = NA_INTEGER;
     } else {
@@ -221,7 +226,7 @@ SEXP vctrs_in(SEXP needles, SEXP haystack) {
   // Load dictionary with haystack
   R_len_t n_haystack = vec_size(haystack);
   for (int i = 0; i < n_haystack; ++i) {
-    uint32_t hash = dict_hash_scalar(&d, &d, i);
+    uint32_t hash = dict_hash_scalar(&d, i);
 
     if (d.key[hash] == DICT_EMPTY) {
       dict_put(&d, hash, i);
@@ -237,7 +242,7 @@ SEXP vctrs_in(SEXP needles, SEXP haystack) {
   int* p_out = LOGICAL(out);
 
   for (int i = 0; i < n_needle; ++i) {
-    uint32_t hash = dict_hash_scalar(&d, &d_needles, i);
+    uint32_t hash = dict_hash_with(&d, &d_needles, i);
     p_out[i] = (d.key[hash] != DICT_EMPTY);
   }
 
@@ -255,7 +260,7 @@ SEXP vctrs_count(SEXP x) {
 
   R_len_t n = vec_size(x);
   for (int i = 0; i < n; ++i) {
-    int32_t hash = dict_hash_scalar(&d, &d, i);
+    int32_t hash = dict_hash_scalar(&d, i);
 
     if (d.key[hash] == DICT_EMPTY) {
       dict_put(&d, hash, i);
@@ -302,7 +307,7 @@ SEXP vctrs_duplicated(SEXP x) {
 
   R_len_t n = vec_size(x);
   for (int i = 0; i < n; ++i) {
-    int32_t hash = dict_hash_scalar(&d, &d, i);
+    int32_t hash = dict_hash_scalar(&d, i);
 
     if (d.key[hash] == DICT_EMPTY) {
       dict_put(&d, hash, i);
@@ -316,7 +321,7 @@ SEXP vctrs_duplicated(SEXP x) {
   int* p_out = LOGICAL(out);
 
   for (int i = 0; i < n; ++i) {
-    int32_t hash = dict_hash_scalar(&d, &d, i);
+    int32_t hash = dict_hash_scalar(&d, i);
     p_out[i] = p_val[hash] != 1;
   }
 
@@ -345,7 +350,7 @@ SEXP vctrs_duplicate_split(SEXP x) {
 
   // Fill dictionary, out_pos, and count
   for (int i = 0; i < n; ++i) {
-    uint32_t hash = dict_hash_scalar(&d, &d, i);
+    uint32_t hash = dict_hash_scalar(&d, i);
 
     if (d.key[hash] == DICT_EMPTY) {
       p_tracker[hash] = d.used;