--- a/src/dictionary.h
+++ b/src/dictionary.h
@@ -17,7 +17,7 @@ struct dictionary {
 };
 typedef struct dictionary dictionary;
 
-void dict_init(dictionary* d, SEXP x, bool hashed);
+void dict_init(dictionary* d, SEXP x);
 void dict_free(dictionary* d);
 uint32_t dict_hash_scalar(dictionary* d, dictionary* x, R_len_t i);
 void dict_put(dictionary* d, uint32_t k, R_len_t i);