--- a/src/dictionary.h
+++ b/src/dictionary.h
@@ -17,9 +17,29 @@ struct dictionary {
 };
 typedef struct dictionary dictionary;
 
+/**
+ * Initialise a dictionary
+ *
+ * - `dict_init()` creates a dictionary and precaches the hashes for
+ *   each element of `x`.
+ *
+ * - `dict_init_partial()` creates a dictionary with precached hashes
+ *   as well, but does not allocate an array of keys. This is useful
+ *   for finding a key in another dictionary with `dict_hash_with()`.
+ */
 void dict_init(dictionary* d, SEXP x);
 void dict_init_partial(dictionary* d, SEXP x);
 void dict_free(dictionary* d);
-uint32_t dict_hash_with(dictionary* d, dictionary* x, R_len_t i);
+
+/**
+ * Find key hash for a vector element
+ *
+ * - `dict_hash_scalar()` returns the key hash for element `i`.
+ *
+ * - `dict_hash_with()` finds the hash for indexing into `d` with
+ *   element `i` of `x`.
+ */
 uint32_t dict_hash_scalar(dictionary* d, R_len_t i);
+uint32_t dict_hash_with(dictionary* d, dictionary* x, R_len_t i);
+
 void dict_put(dictionary* d, uint32_t k, R_len_t i);