… src/hash.c
… src/vctrs.h