This way the hashes can be computed with the vectorised method.

```r
x <- 1:1000

bench::mark(
  before = vec_hash(x, rowwise = TRUE),
  after = vec_hash(x, rowwise = FALSE),
  check = FALSE
)[1:6]
#>   expression      min     mean   median      max `itr/sec`
#>   <chr>      <bch:tm> <bch:tm> <bch:tm> <bch:tm>     <dbl>
#> 1 before       22.8µs   37.2µs   33.8µs   1.16ms    26851.
#> 2 after        11.7µs   17.7µs   13.8µs  888.9µs    56583.


df <- tibble(x = 1:1000, y = 1000:1)

bench::mark(
  before = vec_hash(df, rowwise = TRUE),
  after = vec_hash(df, rowwise = FALSE),
  check = FALSE
)[1:6]
#>   expression      min     mean   median      max `itr/sec`
#>   <chr>      <bch:tm> <bch:tm> <bch:tm> <bch:tm>     <dbl>
#> 1 before       83.6µs  103.7µs   87.3µs    736µs     9645.
#> 2 after        18.8µs   25.7µs   21.6µs    690µs    38915.


nested_df <- tibble(x = 1:1000, y = tibble(z = 1:1000))

bench::mark(
  before = vec_hash(nested_df, rowwise = TRUE),
  after = vec_hash(nested_df, rowwise = FALSE),
  check = FALSE
)[1:6]
#>   expression      min     mean   median      max `itr/sec`
#>   <chr>      <bch:tm> <bch:tm> <bch:tm> <bch:tm>     <dbl>
#> 1 before      140.1µs  174.4µs  149.8µs    998µs     5733.
#> 2 after        18.8µs   28.2µs   22.1µs    701µs    35450.
```

The hashes are also always precomputed, even in `_any` functions that could be short-circuited. This makes the code simpler and the best case performance gain of not prehashing doesn't seem to outweigh the other costs:

```r
worst <- 1:1000
mid <- c(1:500, 500L, 502:1000)
best <- c(1L, 1L, 3:1000)

bench::mark(
  before_worst = vec_duplicate_any(worst),
  before_mid = vec_duplicate_any(mid),
  before_best = vec_duplicate_any(best),
  check = FALSE
)[1:6]
#>   expression        min     mean   median      max `itr/sec`
#>   <chr>        <bch:tm> <bch:tm> <bch:tm> <bch:tm>     <dbl>
#> 1 before_worst  40.27µs  48.21µs  41.06µs    903µs    20742.
#> 2 before_mid    12.37µs  14.96µs  13.02µs    711µs    66861.
#> 3 before_best    2.36µs   4.21µs   3.19µs    845µs   237600.


bench::mark(
  after_worst = vec_duplicate_any(worst),
  after_mid = vec_duplicate_any(mid),
  after_best = vec_duplicate_any(best),
  check = FALSE
)[1:6]
#>   expression       min     mean   median      max `itr/sec`
#>   <chr>       <bch:tm> <bch:tm> <bch:tm> <bch:tm>     <dbl>
#> 1 after_worst     28µs   34.1µs  28.86µs    633µs    29350.
#> 2 after_mid      9.7µs   13.1µs  10.61µs    969µs    76379.
#> 3 after_best    3.48µs    5.5µs   4.23µs    788µs   181721.
```